Add a LDAPError abstraction.
The current implementation doesn't allow for nice error handling using
the try! macro. This change positions us to refactor the RustLDAP
library to return LDAPError, which implement the traits needed to use
try!, among other things. Currently, all errors are naively packed as a
generic error, but it sets us up to do more fine grained error handling,
so users can handle errors intelligently.

Be warned: this major change is a change to the public API of the
library. If we're using semantic versioning, this should be a major + 1.